feat: add select buttons for different user roles in UserTable component
Add select buttons for Guru, MPK, and OSIS roles in the UserTable component. These buttons allow the user to select all users with the corresponding role. This enhances the user experience and improves the usability of the UserTable component.

Also, update the maxAge value in the session configuration of AuthOptions to 24 hours. This ensures that user sessions expire after 24 hours, enhancing security.

Finally, update the submitAcess and deleteAcess functions in voteSession.actions to revalidate specific paths after successful operations. This ensures that the vote and votesesion pages, as well as the admin/votesesion page, are updated with the latest data.

Closes #123